+++ src/components/AssessmentForms/index.js
@@ -1,0 +1,3 @@
+export { default as AttitudeScoreCard } from './AttitudeForm/AttitudeScoreCard';
+export { default as AssessmentScoreCard } from './AssessmentCard/AssessmentScoreCard';
+export { default as ImprovementGraph } from './ImprovementGraph';
