+++ src/pages/assessment-scorecard-input/assessment-scorecard-input.jsx
@@ -1,34 +1,10 @@
 import React, { useState, useEffect } from 'react';
-import PropTypes from 'prop-types';
 import { useParams } from 'react-router-dom';
 import { TLPBackend, scrollToTop } from '../../common/utils';
 
 import ReturnHeader from '../../common/ReturnHeader/ReturnHeader';
-import AssessmentScoreCard from '../../components/AssessmentCard/AssessmentScoreCard';
-import Graph from '../../components/Graph/Graph';
+import { AssessmentScoreCard, ImprovementGraph } from '../../components/AssessmentForms';
 import styles from './assessment-scorecard-input.module.css';
-
-const ImprovementGraph = ({ studentData }) => {
-  const MAX_ASSESSMENT_SCORE = 93;
-  const MAX_ATTITUDE_SCORE = 80;
-
-  const preAttitude =
-    (studentData?.pretestR?.reduce((tot, a) => tot + a, 0) / MAX_ATTITUDE_SCORE) * 100;
-  const postAttitude =
-    (studentData?.posttestR?.reduce((tot, a) => tot + a, 0) / MAX_ATTITUDE_SCORE) * 100;
-  const preAssessment =
-    (studentData?.pretestA?.reduce((tot, a) => tot + a, 0) / MAX_ASSESSMENT_SCORE) * 100;
-  const postAssessment =
-    (studentData?.posttestA?.reduce((tot, a) => tot + a, 0) / MAX_ASSESSMENT_SCORE) * 100;
-
-  return (
-    <Graph
-      xLabels={['Attitudinal', 'Academic']}
-      preData={[preAttitude, preAssessment]}
-      postData={[postAttitude, postAssessment]}
-    />
-  );
-};
 
 const AssessmentScorecardInput = () => {
   const { studentID } = useParams();
@@ -109,15 +85,6 @@ const AssessmentScorecardInput = () => {
       </div>
     </div>
   );
-};
-
-ImprovementGraph.propTypes = {
-  studentData: PropTypes.shape({
-    pretestR: PropTypes.arrayOf(PropTypes.number),
-    posttestR: PropTypes.arrayOf(PropTypes.number),
-    pretestA: PropTypes.arrayOf(PropTypes.number),
-    posttestA: PropTypes.arrayOf(PropTypes.number),
-  }).isRequired,
 };
 
 export default AssessmentScorecardInput;
